restuctured the helpers
